2019.01.15

+++ associate/m2m/indusland_factory_line.js
@@ -19,7 +19,7 @@ var indusland_factory_line = conn.define('indusland_factory_line',
     'allowNull': true,
 },
 'stay':{
-    'type':Sequelize.INTEGER(11),
+    'type':Sequelize.DOUBLE(2),
     'allowNull': true,
 },
 'condition':{
